requirements.txt
babel
crispy-bootstrap5
django ~= 5.0
django-crispy-forms
django-heroku
django-select2
dominate
fontawesomefree
gunicorn
pyyaml
